+++ app/api/api_v1/endpoints/push_notification.py
@@ -16,4 +16,4 @@ async def create_push_notification_subscription(request: Request, db: AsyncIOMot
 
 @router.post("/trigger-push-notifications")
 def push_notifications(notification: PushNotificationSchema, db: AsyncIOMotorClient = Depends(get_database)):
-    return push_notifications_to_users(db, notification)
+    return push_notifications_to_users(notification)
