+++ app/core/utils.py
@@ -6,8 +6,6 @@ from pywebpush import webpush, WebPushException
 from ..config import get_settings
 
 
-
-
 settings = get_settings()
