--- app/push_notification/schemas.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-
-class PushNotificationSchema(BaseModel):
-    title: str 
-    body: str
-
