+++ docker-compose.yml
@@ -13,7 +13,7 @@ services:
         ports:
             - 8000:8000
         env_file: 
-            - ./.environ/.dev-sample
+            - .env
     
     
     celery_worker:
@@ -26,6 +26,6 @@ services:
             - .:/app
     
         env_file: 
-            - ./.environ/.dev-sample
+            - .env
         depends_on:
             - web
